work on kruskal and create UnionFind

+++ src/model/IGraph.java
@@ -3,7 +3,7 @@ package model;
 public interface IGraph <V>{
 	
 	public boolean addVertex(V v);
-	public void addEdge(V u, V v, double w);
+	public void addEdge(V u, V v, int w);
 	public IGraph<V> bfs(V v);
 	public IGraph<V> dfs(V v);
 	public IGraph<V> prim(IGraph<V> g, V v);
